requirements.txt
openai-agents>=0.0.4
python-dotenv>=1.0.1
openai>=1.0.0
pydantic>=2.0.0 